requirements.txt
google-adk
python-dotenv
requests
pydantic
